# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Evolvus Swagger API",
    "description": "Evolvus Spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/entity/{entityCode}": {
      "put": {
        "tags": [
          "Entity API"
        ],
        "summary": "Update an entity to the database",
        "description": "",
        "operationId": "updateEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/entity": {
      "post": {
        "tags": [
          "Entity API"
        ],
        "summary": "Add a new entity to the database",
        "description": "",
        "operationId": "addEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Entity API"
        ],
        "summary": "Get an entity from the database.",
        "description": "",
        "operationId": "getEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be get from the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "api/lookup": {
      "post": {
        "tags": [
          "Lookup API"
        ],
        "summary": "Add a new lookup to the database",
        "description": "",
        "operationId": "addlookup",
        "consumes": [
          "lookup/json"
        ],
        "produces": [
          "lookup/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "lookup object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lookup"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Lookup API"
        ],
        "summary": "Get an lookup from the database.",
        "description": "",
        "operationId": "getlookup",
        "consumes": [
          "lookup/json"
        ],
        "produces": [
          "lookup/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "Entity": {
      "type": "object",
      "required": [
        "name",
        "entityCode",
        "tenantId",
        "parent"
      ],
      "tags": [
        {
          "name": "string",
          "entityCode": "string",
          "description": "string",
          "tenantId": "string",
          "parent": "string"
        }
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "MGRoad"
        },
        "entityCode": {
          "type": "string",
          "example": "MGRoad"
        },
        "description": {
          "type": "string",
          "example": "This is MGRoad Entity"
        },
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "parent": {
          "type": "string",
          "example": "HEADOFFICET002"
        }
      }
    },
...
    "lookup": {
      "type": "object",
      "required": [
        "lookupCode",
        "value",
        "tenantId"
      ],
      "tags": [
        {
          "tenantId": "string",
          "lookupCode": "string",
          "value": "string",
          "valueOne": "string",
          "valueTwo": "string",
          "valueThree": "string",
          "valueFour": "string",
          "valueFive": "string",
          "valueSix": "string",
          "valueSeven": "string",
          "valueEight": "string",
          "valueNine": "string",
          "valueTen": "string"
        }
      ],
      "properties": {
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "lookupCode": {
          "type": "string",
          "example": "PRODUCT_CODE"
        },
        "value": {
          "type": "string",
          "example": "PENDING_AUTHORIZATION"
        },
        "valueOne": {
          "type": "string",
          "example": "MANDATE"
        },
        "valueTwo": {
          "type": "string",
          "example": "zip"
        },
        "valueThree": {
          "type": "string",
          "example": "BULK_MANDATE"
        },
        "valueFour": {
          "type": "string",
          "example": "ORG"
        },
        "valueFive": {
          "type": "string",
          "example": "ACH_TR"
        },
        "valueSix": {
          "type": "string",
          "example": "ORG/ACHCR"
        },
        "valueSeven": {
          "type": "string",
          "example": "txt"
        },
        "valueEight": {
          "type": "string",
          "example": "CSV"
        },
        "valueNine": {
          "type": "string",
          "example": "ASPCR"
        },
        "valueTen": {
          "type": "string",
          "example": "EMBK"
        }
      }
    }
  }
}